Fix debug enabled games to behave exactly as non-debugged
In the code that checked for debug a para() was made if there was
no debugging info. This caused the first output in a debug enabled
game to be different if it was compiled without.

This para() was moved to a common place which caused many tests to
differ with this single empty line.