Move checking of command line options vs. source to analyze()
Doing this in generate() was to late since generation of source tables
are done in analyze phase. Earlier that was done if -debug was used on
command line. But since we are now unifying that with the source
options, hopefully this will be consistent and we can simply use opts[].