[build][gargoyle] Do not build garglk to link gararun
Previously we leveraged the garglk build (which used Jam). Since it is
always a hazzle to setup up compilation of other systems (libraries
and all) and garglk have moved away from Jam to CMake it seemed easier
to just compile and link locally since we only use the .exe anyway.

It does require a libgarglk.dll to link against so that is provided in
the repo. This will work as long as the interface to it does not change.